+++ docs/guide/changelog.md
@@ -8,6 +8,8 @@
 
 - 新增 [颜色选择器 ColorPicker](https://themusecatcher.github.io/vue-amazing-ui/guide/components/colorpicker.html) 组件
 - 优化并更新 [滑动输入条 Slider](https://themusecatcher.github.io/vue-amazing-ui/guide/components/slider.html) 组件代码逻辑
+- 优化并更新 [滚动条 Scrollbar](https://themusecatcher.github.io/vue-amazing-ui/guide/components/scrollbar.html) 组件代码逻辑
+- 优化并更新 [图片 Image](https://themusecatcher.github.io/vue-amazing-ui/guide/components/image.html) 组件代码，新增 draggable 属性，支持设置是否可以拖动图片
 - 组件库及文档代码优化
 
 ## 2.0.9 <Tag color="volcano" size="small">2025-01-10</Tag>